pyproject.toml
[project]
name = "target-kafka"
version = "0.0.1"
description = "Singer target for Kafka, built with the Meltano Singer SDK."
readme = "README.md"
authors = [{ name = "Rishi Pradeepkumar", email = "rishi.pradeepkumar@think41.com" }]
keywords = [
    "ELT",
    "Kafka",
]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
license = "Apache-2.0"
license-files = [ "LICENSE" ]
requires-python = ">=3.10"
dependencies = [
    "singer-sdk~=0.53.5",
    "requests~=2.32.3",
    "kafka-python>=2.0.2",
    "python-snappy"
]

[project.optional-dependencies]
s3 = [
    "s3fs>=2026.1.0",
]

[project.scripts]
# CLI declaration
target-kafka = 'target_kafka.target:TargetKafka.cli'

[dependency-groups]
dev = [
    { include-group = "test" },
]
test = [
    "pytest>=9",
    "pytest-github-actions-annotate-failures>=0.3",
    "singer-sdk[testing]",
]
typing = [
    "mypy>=1.16.0",
    "ty>=0.0.1-alpha.16",
    "types-requests",
]

[tool.pytest]
addopts = [
    "--durations=10",
    "-ra",
    "--strict-config",
    "--strict-markers",
]
filterwarnings = [
    "error",
    "once:No records were available to test:UserWarning",
]
log_level = "INFO"
minversion = "9.0"
testpaths = [
    "tests",
]
xfail_strict = true

[tool.mypy]
enable_error_code = [
    "ignore-without-code",
    "redundant-expr",
    "truthy-bool",
]
warn_redundant_casts = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.ruff]
line-length = 100
required-version = ">=0.14"

[tool.ruff.lint]
future-annotations = true
ignore = [
    "COM812",  # missing-trailing-comma
]
select = ["ALL"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = [
    "hatchling>=1,<2",
]
build-backend = "hatchling.build"

# This configuration can be used to customize tox environments
[tool.tox]
min_version = "4.32"
requires = [
    "tox>=4.32",
    "tox-uv",
]
env_list = [
    "typing",
    "py314",
    "py313",
    "py312",
    "py311",
    "py310",
]

[tool.tox.env_run_base]
runner = "uv-venv-lock-runner"
pass_env = [
    "GITHUB_*",
    "TARGET_KAFKA_*",
]
dependency_groups = [ "test" ]
commands = [ [ "pytest", { replace = "posargs", default = [ "tests" ], extend = true } ] ]

[tool.tox.env.typing]
dependency_groups = [ "test", "typing" ]
commands = [
    [ "mypy", { replace = "posargs", default = [ "target_kafka", "tests" ], extend = true } ],
    [ "ty", "check", { replace = "posargs", default = [ "target_kafka", "tests" ], extend = true } ],
]
